feat: Support infallible conversion
In some cases, it may be desireable to generate values that are known to
be valid. Rather than having to use `try_from` and `unwrap` (since we
know the value is valid), there's now an infallible `from` method that
can be used when `Check::Err` is `Infallible`.

Note that this would be unncessary with the 'never' type (`!`) and
methods on `Result` to unwrap infallibly when the `Err` type is never.
These are both currently unstable, however.