feat: Add `serde` support behind a feature
This makes it possible to directly serialize `Checked<T>` values, which
will be transparently serialized as if it were just `T`. `Checked<T>`
values can also be deserialized, but only if `T`'s `Check` impl returns
itself as the OK variant (otherwise, we would not know what to
deserialize first).